+++ Modules/Microsoft365DSC/DscResources/MSFT_EXOIntraOrganizationConnector/MSFT_EXOIntraOrganizationConnector.psm1
@@ -102,14 +102,14 @@ function Get-TargetResource
 
         Write-Verbose -Message "Found IntraOrganizationConnector $($Identity)"
 
-        $DiscoveryEndpointValue = $IntraOrganizationConnector.DiscoveryEndpoint.ToString()
+        $DiscoveryEndpointValue = $IntraOrganizationConnector.DiscoveryEndpoint
         if (-not $DiscoveryEndpointValue.EndsWith('/'))
         {
             $DiscoveryEndpointValue += '/'
         }
         if ($IntraOrganizationConnector.TargetSharingEpr)
         {
-            $TargetSharingEprValue = $IntraOrganizationConnector.TargetSharingEpr.AbsoluteUri.ToString()
+            $TargetSharingEprValue = $IntraOrganizationConnector.TargetSharingEpr
         }
         else
         {